Commit before merge New controls

+++ src/robotModules/robotJoint.java
@@ -3,8 +3,8 @@ package robotModules;
 public class robotJoint {
 	
 	
-	private int x;
-	private int y;
+	private double x;
+	private double y;
 	private int globalID;
 	private int currentBentDegree;
 	private int robotBentDegree;
@@ -49,22 +49,22 @@ public class robotJoint {
 	}
 
 
-	public int getX() {
+	public double getX() {
 		return x;
 	}
 
 
-	public void setX(int x) {
+	public void setX(double x) {
 		this.x = x;
 	}
 
 
-	public int getY() {
+	public double getY() {
 		return y;
 	}
 
 
-	public void setY(int y) {
+	public void setY(double y) {
 		this.y = y;
 	}
